tia/dua-tia-ml-model/requirements.txt
adapters==1.0.0
huggingface_hub==0.24.7
imbalanced-learn==0.12.4
pandas==2.2.1
psutil==6.0.0
pyyaml==6.0.2
scikit-learn==1.4.1.post1
torch==2.2.1 --index-url https://download.pytorch.org/whl/cpu
transformers==4.43.4